backend/app/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
python-multipart==0.0.6
pydantic==1.10.7
sqlmodel==0.0.8
psycopg2-binary==2.9.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pillow==10.0.1
torch==2.2.0          # or change to the version you trained with; optional if using TF
# If your model uses torchvision:
torchvision==0.17.0
aiofiles==23.1.0
pytest==7.4.0
httpx==0.25.0
python-dotenv==1.0.0
